Improve yay-install.sh and add podman

+++ scripts/bootstrap/yay-install.sh
@@ -1,8 +1,12 @@
 #!/usr/bin/env bash
 
-pacman -S --needed git base-devel
-git clone https://aur.archlinux.org/yay.git
-cd yay
-makepkg -si
-cd ..
-rm -rf yay
+if [ ! command -v yay &> /dev/null ]; then
+    pacman -S --needed git base-devel
+    git clone https://aur.archlinux.org/yay.git
+    cd yay
+    makepkg -si
+    cd ..
+    rm -rf yay
+else
+    echo "Yay already installed, skipping..."
+fi
